More variables correctness for Iss
git-svn-id: https://www.soclib.fr/svn/trunk/soclib@1450 2cf4c1f0-a32b-0410-9068-d80f8be93c97

+++ soclib/iss/arm/include/arm.h
@@ -103,7 +103,7 @@ public:
     static const unsigned int s_pc_register_no = 15;
     static const Iss2::debugCpuEndianness s_endianness = Iss2::ISS_LITTLE_ENDIAN;
 
-    static const unsigned int n_irq = 2;
+    static const size_t n_irq = 2;
 
     void setICacheInfo( size_t line_size, size_t assoc, size_t n_lines );
     void setDCacheInfo( size_t line_size, size_t assoc, size_t n_lines );
